proof_of_concept/03_pipelines/requirements.txt
confluent-kafka
prometheus_client
cassandra-driver
pyyaml

# AUDITED 03/09 -- OK
# pip install -r requirements.txt